+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Instances.hs
@@ -70,7 +70,7 @@ import Test.Cardano.Ledger.Shelley.Utils (epochFromSlotNo)
 
 shelleyGenPParams :: EraPParams era => Gen (PParams era)
 shelleyGenPParams = do
-  minfeeA <- Coin <$> choose (0, 1000)
+  minfeeA <- CoinPerByte . Coin <$> choose (0, 1000)
   minfeeB <- Coin <$> choose (0, 10000)
   pure $
     emptyPParams