+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## 1.15.0.0
 
+* Changed type of `appMinFeeA` to `CoinPerByte`
 * Change sets containing errors into `NonEmptySet` for `AlonzoUtxoPredFailure`, `AlonzoUtxowPredFailure`
 * Change all lists into `NonEmpty` for `AlonzoUtxoPredFailure`, `AlonzoUtxosPredFailure`, `AlonzoUtxowPredFailure`
 * Change `collectPlutusScriptsWithContext` to return a `NonEmpty` list of failures
 * Changed the type of `dappMinUTxOValue` to `CompactForm Coin` in `DowngradeAlonzoPParams`
 * Changed the type of the following fields to `CompactForm Coin` in `AlonzoPParams`:
-  - `appMinFeeA`
   - `appMinFeeB`
   - `appKeyDeposit`
   - `appMinPoolCost`