+++ eras/babbage/impl/CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## 1.13.0.0
 
+* Changed type of `bppMinFeeA` to `CoinPerByte`
 * Moved `CoinPerByte` to `cardano-ledger-core`
 * Change sets containing errors into `NonEmptySet` for `BabbageUtxowPredFailure`
 * Change all lists into `NonEmpty` for `BabbageUtxoPredFailure`
 * Add `babbageUtxoValidation`
 * Add `babbageUtxoTests`
 * Changed the type of the following fields to `CompactForm Coin` in `BabbagePParams`:
-  - `bppMinFeeA`
   - `bppMinFeeB`
   - `bppKeyDeposit`
   - `bppMinPoolCost`